Implement lzc_list()
This implements a stable zfs list API. inspired by the CLI `zfs list`.
It unifies several existing API calls into a single call that provides
an almost complete subset of the functionality provided by `zfs list`.
The existing API calls that this supercedes are:

	- ZFS_IOC_OBJSET_STATS
	- ZFS_IOC_OBJSET_ZPLPROPS
	- ZFS_IOC_DATASET_LIST_NEXT
	- ZFS_IOC_SNAPSHOT_LIST_NEXT
	- ZFS_IOC_GET_BOOKMARKS (stable lzc API)

Since ZFS_IOC_GET_BOOKMARKS is a stable API, it has been left untouched.

The functionality absent from the API that is present in `zfs list` is
the ability to sort and request data on specific properties. Under this
API, userland is (still) expected to sort and all property data is
returned.

Programming documentation on the API is included a comment in
libzfs_core.c.

Signed-off-by: Richard Yao <ryao@gentoo.org>