Improve zfs_stable_ioc_set_props before it is stable
lzc_set_props() will now expose opts nvlist, provide an errlist and
support atomic operations. THe latter of those requires additional
explanation.

The legacy API was atomic on success and non-atomic on failure such that
a best effort was made to set as many properties as possible and the
stable API initially inherited this behavior. Albert Lee pointed out
that this could race with zfs_list and upon review, this also conflicts
with the atomicity goals for the stable API, so we change it to be
atomic such that the failure to set any property on a dataset will
result in the failure of all properties, rather than just one.

The stable API permits aggregration of calls to set properties on
multiple datasets into one and the guarentee of atomicity is only made
at the dataset level, such that the failure to set properties on 1
dataset will not affect the others. It is not clear that stronger
atomicity is required and the locking required to do it is expensive
enough that this implementation does not attempt it.

Signed-off-by: Richard Yao <ryao@gentoo.org>